fix: changed creating server
